refactor: 디렉터리 분리
