+++ src/main/java/menu/domain/initialdata/InitialMenu.java
@@ -1,8 +1,7 @@
-package menu.domain;
+package menu.domain.initialdata;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public enum InitialMenu {
     JAPANESE(1, "일식", List.of("규동", "우동", "미소시루", "스시", "가츠동", "오니기리", "하이라이스", "라멘", "오코노미야끼")),
